pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pydantic-another-config"
version = "0.1.3"
authors = [
    { name="Evgeniy Mazalov", email="guyboe@gmail.com" },
]
description = "Extends pydantic settings module"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic>=2",
    "python-benedict[yaml]",
]

[tool.setuptools.package-dir]
pydantic_config = "src/pydantic_config"

[project.urls]
"Homepage" = "https://github.com/guyboe/pydantic-config"
"Bug Tracker" = "https://github.com/guyboe/pydantic-config/issues"
